Download sample data to depict the use of the MoveR package.

Given the id of a sample dataset stored in MoveR_SampleData Github repository, this function download the selected sample dataset used as example to run the functions from the MoveR package and returns the local path to access the data.

Usage

DLsampleData(
  dataSet = c(1, 2),
  tracker = c("Ctrax", "IdTracker", "TrackR", "TRex"),
  dir = NULL
)

Arguments

dataSet

A numeric value, either 1 or 2 corresponding to the desired sampled dataset.

tracker

A character string corresponding to the name of the tracking software from which data have to be downloaded: Ctrax, IdTracker, TrackR and/or TRex. By default the function returns the path to sample data from each tracking software.

dir

The path to the directory where the sample dataset is saved, by default, the function will save it in a temporary directory (optional).

Value

This function returns a vector containing the path to the selected sample dataset(s) (raw tracking data), the distance matrix from the border of the arena (color tresholding - greyscale - performed using ImageJ), the reference data (a .csv file containing particles location -manually detected- over several frames to perform sensitivity analyses) and the filtered/cleaned version of the tracking data (filtered using the approach described in the MoveR-Clean-FilterData vignette). In case the sample dataset has already been downloaded in a temporary (default) or specified directory, the function retrieve the data instead of downloading it again.

Author

Quentin PETITJEAN

Examples

## Not run:

# Here we are selecting the first sample dataset, that have been tracked using \href{https://trex.run}{TRex}

DLsampleData(dataSet = 1, tracker = "TRex")
#> [1] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\TRexOutput"                                                            
#> [2] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\ReferenceData\\DistMatrixFromArenaEdge_2602_ISA3080_Low_5.mov_1800.txt"
#> [3] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\ReferenceData\\RefDat_2602_ISA3080_Low_5.csv"                          

# If the "dir" argument is not specified, the data are saved in a temporary directory (deleted after the R session is properly terminated).
# In case the sample data has already been downloaded, the function will retrieve it instead of downloading it again.
# This behavior is similar if the "dir" argument is specified.

## End(Not run)