Apply a custom filter on tracklets.

Given a list of data frame containing tracking information for each tracklet and the result of a specified condition' test as returned by filterFunc this function remove the values that do not meet the condition of the test and split the tracklets accordingly. Also by specifying an additional "minDur" argument, filtered tracklets that have a low number of records can be removed. The function hence returns two sublist, the first containing a summary of the information about the tracklets before and after filtering and the second containing a list with the filtered tracklets according to the condition test and "minDur" argument.

Usage

filterTracklets(trackDat, filter, splitCond = TRUE, minDur = 1, progess = TRUE)

Arguments

trackDat

A list of data frames containing tracking information for each tracklet

filter

A list of vector as returned by filterFunc and containing the result of a condition test.

splitCond

The result of the condition test for which values have to be removed and tracklets have to be splitted (Default: TRUE).

minDur

The minimum duration (i.e., number of records) to consider a tracklet valid after the filtering (default = 1).

progess

A logical value (i.e., TRUE or FALSE) indicating whether a progress bar should be displayed to inform process progression (default = TRUE).

Value

This function returns a list containing two sublists:

"SummaryFiltering": a list containing a summary of the informations about the tracklets before and after filtering:

"Tracknb_before_filter": the initial number of tracklets.

"Tracknb_after_filter": the number of remaining tracklet after filtering, according to the "filter" argument.

"Tracknb_after_minDur": the number of remaining tracklet after filtering that have a number of records above "minDur" argument.

"TotTrackDuration_before_filter": the sum of the records belonging to each tracklet before the filtering.

"TotTrackDuration_after_filter": the sum of the records belonging to each tracklet after the filtering.

"TotTrackDuration_after_minDur": the sum of the records belonging to each tracklet that have a number of records above "minDur" argument after the filtering.

"%Data_kept_after_filter": the percent of records remaining afer the filtering.

"%Data_kept_after_minDur": the percent of records remaining afer the filtering for the tracklets that have a number of records above "minDur" argument only.

"CleanedTracklets": a list of data frames containing the filtered tracklets according to the condition test specified by filterFunc and "minDur" argument.

See also

Author

Quentin PETITJEAN

Examples

## Not run:

# Download the first dataset from the sample data repository
Path2Data <- MoveR::DLsampleData(dataSet = 1, tracker = "TRex")
Path2Data
#> [1] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\TRexOutput"                                                            
#> [2] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\ReferenceData\\DistMatrixFromArenaEdge_2602_ISA3080_Low_5.mov_1800.txt"
#> [3] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\ReferenceData\\RefDat_2602_ISA3080_Low_5.csv"                          

# Import the list containing the 9 vectors classically used for further computation
Data <- MoveR::readTrex(Path2Data[[1]])

# convert it to a list of tracklets
trackDat <- MoveR::convert2Tracklets(Data[1:7], by = "identity")

# example 1:
## test for the presence of infinite value in x.pos, if infinite values are detected, the result is TRUE
FiltInf <- MoveR::filterFunc(trackDat, toFilter = "x.pos", customFunc = function(x) is.infinite(x))

## Then remove infinite values and split the tracklets when an infinite value is found
## here we keep every remaining tracklet, whatever its duration (the number of record within each tracklet)
trackDatNoInf <- MoveR::filterTracklets(trackDat, filter = FiltInf, splitCond = TRUE, minDur = 1)

## Check the summary of the filtering
str(trackDatNoInf$SummaryFiltering)
#> List of 8
#>  $ Tracknb_before_filter         : int 13
#>  $ Tracknb_after_filter          : int 1938
#>  $ Tracknb_after_minDur          : int 1938
#>  $ TotTrackDuration_before_filter: int 156291
#>  $ TotTrackDuration_after_filter : int 141606
#>  $ TotTrackDuration_after_minDur : int 141606
#>  $ %Data_kept_after_filter       : num 90.6
#>  $ %Data_kept_after_minDur       : num 90.6

## take a look at the first filtered tracklet
str(trackDatNoInf$CleanedTracklets[[1]])
#> 'data.frame':	17 obs. of  7 variables:
#>  $ maj.ax  : num  23.5 23.9 21.6 23.2 21.5 ...
#>  $ angle   : num  -0.733 -0.767 -0.761 -0.896 -0.88 ...
#>  $ min.ax  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ x.pos   : num  2278 2273 2269 2268 2267 ...
#>  $ y.pos   : num  633 638 641 645 646 ...
#>  $ identity: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ frame   : num  0 1 2 3 4 5 6 7 8 9 ...

## alternatively, we can only keep the tracklets with a duration above 10 (the number of record within each tracklet)
## by modifying the minDur argument
trackDatNoInfDur10 <- MoveR::filterTracklets(trackDat, filter = FiltInf, splitCond = TRUE, minDur = 10)

## Check the summary of the filtering
str(trackDatNoInfDur10$SummaryFiltering)
#> List of 8
#>  $ Tracknb_before_filter         : int 13
#>  $ Tracknb_after_filter          : int 1938
#>  $ Tracknb_after_minDur          : int 790
#>  $ TotTrackDuration_before_filter: int 156291
#>  $ TotTrackDuration_after_filter : int 141606
#>  $ TotTrackDuration_after_minDur : int 139053
#>  $ %Data_kept_after_filter       : num 90.6
#>  $ %Data_kept_after_minDur       : num 89

## take a look at the first filtered tracklet
str(trackDatNoInfDur10$CleanedTracklets[[1]])
#> 'data.frame':	17 obs. of  7 variables:
#>  $ maj.ax  : num  23.5 23.9 21.6 23.2 21.5 ...
#>  $ angle   : num  -0.733 -0.767 -0.761 -0.896 -0.88 ...
#>  $ min.ax  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ x.pos   : num  2278 2273 2269 2268 2267 ...
#>  $ y.pos   : num  633 638 641 645 646 ...
#>  $ identity: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ frame   : num  0 1 2 3 4 5 6 7 8 9 ...

# example 2: test for the length of the particles:
## keep only particles with a size included between 1 and 20 pixels
## if particles' size is ranging between 1 and 20 pixels, the result is TRUE.
FiltSize <- MoveR::filterFunc(trackDat, toFilter = "maj.ax", customFunc = function(x) x >= 1 & x <= 20)

## Then remove values that are not included within the [1:20] interval and split the tracklets accordingly
## here we keep every remaining tracklet, whatever its duration (the number of record within each tracklet)
trackDatSize20 <- MoveR::filterTracklets(trackDat, filter = FiltSize, splitCond = TRUE, minDur = 1)

## Check the summary of the filtering
str(trackDatSize20$SummaryFiltering)
#> List of 8
#>  $ Tracknb_before_filter         : int 13
#>  $ Tracknb_after_filter          : int 15534
#>  $ Tracknb_after_minDur          : int 15534
#>  $ TotTrackDuration_before_filter: int 156291
#>  $ TotTrackDuration_after_filter : int 125592
#>  $ TotTrackDuration_after_minDur : int 125592
#>  $ %Data_kept_after_filter       : num 80.4
#>  $ %Data_kept_after_minDur       : num 80.4

## take a look at the first filtered tracklet
str(trackDatSize20$CleanedTracklets[[1]])
#> 'data.frame':	5 obs. of  7 variables:
#>  $ maj.ax  : num  23.5 23.9 21.6 23.2 21.5
#>  $ angle   : num  -0.733 -0.767 -0.761 -0.896 -0.88
#>  $ min.ax  : num  NA NA NA NA NA
#>  $ x.pos   : num  2278 2273 2269 2268 2267
#>  $ y.pos   : num  633 638 641 645 646
#>  $ identity: num  0 0 0 0 0
#>  $ frame   : num  0 1 2 3 4

## alternatively, we can only keep the tracklets with a duration above 10 (the number of record within each tracklet) by modifying the minDur argument
trackDatSize20Dur10 <- MoveR::filterTracklets(trackDat, filter = FiltSize, splitCond = TRUE, minDur = 10)

## Check the summary of the filtering
str(trackDatSize20Dur10$SummaryFiltering)
#> List of 8
#>  $ Tracknb_before_filter         : int 13
#>  $ Tracknb_after_filter          : int 15534
#>  $ Tracknb_after_minDur          : int 3256
#>  $ TotTrackDuration_before_filter: int 156291
#>  $ TotTrackDuration_after_filter : int 125592
#>  $ TotTrackDuration_after_minDur : int 91523
#>  $ %Data_kept_after_filter       : num 80.4
#>  $ %Data_kept_after_minDur       : num 58.6

## take a look at the first filtered tracklet
str(trackDatSize20Dur10$CleanedTracklets[[1]])
#> 'data.frame':	22 obs. of  7 variables:
#>  $ maj.ax  : num  25 29.2 26.6 23.3 21.6 ...
#>  $ angle   : num  0.734 -2.527 -2.611 -2.563 0.36 ...
#>  $ min.ax  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ x.pos   : num  2470 2471 2470 2469 2469 ...
#>  $ y.pos   : num  756 757 758 757 757 ...
#>  $ identity: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ frame   : num  59 60 61 62 63 64 65 66 67 68 ...

## End(Not run)