Get information in a tracklets object.

Given an object of class "tracklets", this function returns some information about tracking processing from the tracklets object.

Usage

getInfo(trackDat, attrName = NULL)

Arguments

trackDat

An object of class "tracklets" containing a list of tracklets and their characteristics classically used for further computations (at least x.pos, y.pos, frame).

attrName

A character string indicating which attirbute to return. If this argument is not provided the function returns a list of all hidden attributes (default = NULL).

Value

The value of the selected attribute or a list of the attributes' values in the tracklets object. By default (attrName = NULL), the function returns a list of all hidden attributes (frameR, scale, imgRes).

See also

Author

Quentin PETITJEAN

Examples

if (FALSE) {

# Download the first dataset from the sample data repository
Path2Data <- MoveR::DLsampleData(dataSet = 1, tracker = "TRex")
Path2Data

# Import the data as an object of class "tracklets"
# and flip Y coordinates to start on the bottom-left
Data <- MoveR::readTrex(Path2Data[[1]],
               flipY = T,
               imgHeight = 2160,
               rawDat = F
        )
        
str(Data)

# get the tracking informations from the tracklets object (frameR corresponding to the frame rate is already returned because it is included in TRex output)
getInfo(Data)

# set the image resolution in the tracklets object
Data <- setInfo(Data, imgRes = c(3840, 2160))

# check the result (now the tracklets object contains both information about the frame rate and the image resolution, which can be used in further functions of the MoveR package automatically - without specifying it in function's argument)
getInfo(Data)

}