Import Trex tracking output .npz files.

Given the path of a folder containing multiple Trex outputs (one .npz file per particle/individual), this function returns a list of 9 vectors classically used for further computations using MoveR package:

'maj.ax': the length of the major axis (i.e., the midline) for a particle over frame (i.e., length of the ellipse).

'angle': the particle's absolute angle in radians, orientation of the particle according to X-axis.

'min.ax': the length of the minor axis for a particle over frame (i.e., width of the ellipse), returns NA since it is not present in the TREX output.

'x.pos': x position of the particle's centroid.

'y.pos': y position of the particle's centroid.

'identity': the particle's identity given by the tracking software.

'frame': the video frame number at which the measurements has been made.

'ntargets': the number of particle tracked over each frame.

'timestamps': the elapsed time over each frame, in seconds.

Alternatively, the function can returns a list containing 2 sublists, the first corresponding to the one mentioned above and the second containing all the elements retrieved from the .npz files (see rawDat argument).

Usage

readTrex(trexPath, flipY = FALSE, imgHeight = NULL, rawDat = FALSE)

Arguments

trexPath

The path of the TRex output folder where .npz files are stored.

flipY

A logical value (i.e., TRUE or FALSE) indicating whether the origin of y coordinates should be flipped. If TRUE, y coordinates are flipped to start on the bottom-left (default = FALSE).

imgHeight

A numeric value expressed in pixels, the length of Y axis corresponding to the height of the image or video resolution (optional, only used when flipY = TRUE).

rawDat

TRUE or FALSE, if TRUE add a second list containing all the elements retrieved from .npz files (see trex.run), may drastically increase the size of the object returned by the function (default = FALSE).

Value

A list containing either a list of 9 elements classically used for further computations or a list containing 2 sublists, the first corresponding to the one previously mentioned and the second containing all the elements retrieved from the .npz files (see rawDat argument). Also, by default the function flip y coordinates to start on the bottom-left.

References

Tristan Walter, Iain D Couzin (2021) TRex, a fast multi-animal tracking system with markerless identification, and 2D estimation of posture and visual fields eLife 10:e64000. trex.run

See also

Author

Quentin PETITJEAN

Examples

## Not run:

# Download the first dataset from the sample data repository
Path2Data <- MoveR::DLsampleData(dataSet = 1, tracker = "TRex")
Path2Data
#> [1] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\TRexOutput"                                                            
#> [2] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\ReferenceData\\DistMatrixFromArenaEdge_2602_ISA3080_Low_5.mov_1800.txt"
#> [3] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpIBGA5h\\MoveR_SampleData-main\\sample_1\\ReferenceData\\RefDat_2602_ISA3080_Low_5.csv"                          

# Import the list containing the 9 vectors classically used for further computation
# and flip Y coordinates to start on the bottom-left
Data <- MoveR::readTrex(Path2Data[[1]],
               flipY = T,
               imgHeight = 2160,
               rawDat = F
        )
str(Data)
#> List of 9
#>  $ maj.ax    : num [1:156291] 23.5 23.9 21.6 23.2 21.5 ...
#>  $ angle     : num [1:156291] -0.733 -0.767 -0.761 -0.896 -0.88 ...
#>  $ min.ax    : num [1:156291] NA NA NA NA NA NA NA NA NA NA ...
#>  $ x.pos     : num [1:156291] 2278 2273 2269 2268 2267 ...
#>  $ y.pos     : num [1:156291] 1527 1522 1519 1515 1514 ...
#>  $ identity  : num [1:156291] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ frame     : num [1:156291] 0 1 2 3 4 5 6 7 8 9 ...
#>  $ ntargets  : num [1:12024] 13 10 8 8 8 8 8 8 8 7 ...
#>  $ timestamps: num [1:12024] 0 0.04 0.08 0.12 0.16 0.2 0.24 0.28 0.32 0.36 ...

# Import the list containing 2 sublists, the first containing the 9 vectors classically used for further computation
# and the second list containing all the elements retrieved from .npz files,
# also do not flip Y coordinates (start on the top-left)
DataFull <- MoveR::readTrex(Path2Data[[1]],
                   rawDat = T
            )
str(DataFull)
#> List of 2
#>  $ Data_Trex    :List of 9
#>   ..$ maj.ax    : num [1:156291] 23.5 23.9 21.6 23.2 21.5 ...
#>   ..$ angle     : num [1:156291] -0.733 -0.767 -0.761 -0.896 -0.88 ...
#>   ..$ min.ax    : num [1:156291] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ x.pos     : num [1:156291] 2278 2273 2269 2268 2267 ...
#>   ..$ y.pos     : num [1:156291] 633 638 641 645 646 ...
#>   ..$ identity  : num [1:156291] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ frame     : num [1:156291] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ ntargets  : num [1:12024] 13 10 8 8 8 8 8 8 8 7 ...
#>   ..$ timestamps: num [1:12024] 0 0.04 0.08 0.12 0.16 0.2 0.24 0.28 0.32 0.36 ...
#>  $ Data_Trex_Raw:List of 31
#>   ..$ BORDER_DISTANCE#pcentroid: num [1:156291] 634 640 644 646 648 ...
#>   ..$ ACCELERATION#pcentroid   : num [1:156291] 0 5716 1187 2206 139 ...
#>   ..$ ACCELERATION#wcentroid   : num [1:156291] 0 4371 1425 2476 1215 ...
#>   ..$ SPEED#smooth#wcentroid   : num [1:156291] 0 174.9 142.4 87.5 39.4 ...
#>   ..$ ANGULAR_A#centroid       : num [1:156291] 0 4001.2 -4063.9 58 -68.1 ...
#>   ..$ ANGULAR_V#centroid       : num [1:156291] 0 160.049 -2.506 -0.185 -2.909 ...
#>   ..$ normalized_midline       : num [1:156291] -0.0717 0.145 0.1258 0.1932 -0.066 ...
#>   ..$ SPEED#wcentroid          : num [1:156291] 0 174.9 142.4 87.5 39.4 ...
#>   ..$ SPEED#pcentroid          : num [1:156291] 0 228.6 212.1 131.1 83.4 ...
#>   ..$ MIDLINE_OFFSET           : num [1:156291] -0.0801 0.1628 0.1041 0.1977 -0.0543 ...
#>   ..$ midline_length           : num [1:156291] 23.5 23.9 21.6 23.2 21.5 ...
#>   ..$ segment_length           : num [1:156291] 2.13 2.17 1.96 2.11 1.96 ...
#>   ..$ frame_segments           : num [1:7578] 0 16 29 30 31 ...
#>   ..$ segment_vxys             : num [1:551268] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ X#wcentroid              : num [1:156291] 2278 2273 2269 2268 2267 ...
#>   ..$ Y#wcentroid              : num [1:156291] 633 638 641 645 646 ...
#>   ..$ num_pixels               : num [1:156291] 173 171 174 186 169 163 172 196 193 205 ...
#>   ..$ midline_x                : num [1:156291] 2286 2281 2275 2275 2274 ...
#>   ..$ midline_y                : num [1:156291] 626 632 635 637 639 ...
#>   ..$ timestamp                : num [1:156291] 0 40000 80000 120000 160000 200000 240000 280000 320000 360000 ...
#>   ..$ missing                  : num [1:156291] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ ANGLE                    : num [1:156291] -0.733 -0.767 -0.761 -0.896 -0.88 ...
#>   ..$ SPEED                    : num [1:156291] 0 43.2 39 132.8 153.4 ...
#>   ..$ frame                    : num [1:156291] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ time                     : num [1:156291] 0 0.04 0.08 0.12 0.16 ...
#>   ..$ AX                       : num [1:156291] 0 997 3282 6059 3629 ...
#>   ..$ AY                       : num [1:156291] 0 414 -4185 -3220 -4406 ...
#>   ..$ VX                       : num [1:156291] 0 39.9 16.9 121.7 123.1 ...
#>   ..$ VY                       : num [1:156291] 0 16.6 -35.1 -53.2 -91.5 ...
#>   ..$ X                        : num [1:156291] 2285 2280 2274 2273 2273 ...
#>   ..$ Y                        : num [1:156291] 628 634 637 639 641 ...

## End(Not run)