Import TrackR tracking output .csv file.

Given the path of the .csv file corresponding to TrackR output, this function returns a list of 9 vectors classically used for further computations using MoveR package:

'maj.ax': the length of the major axis (i.e., the midline) for a particle over frame (i.e., length of the ellipse).

'angle': the particle's absolute angle in radians, orientation of the particle according to y-axis.

'min.ax': the length of the minor axis for a particle over frame (i.e., width of the ellipse).

'x.pos': x position of the particle's centroid.

'y.pos': y position of the particle's centroid.

'identity': the particle's identity given by the tracking software.

'frame': the video frame number at which the measurements has been made.

'ntargets': the number of particle tracked over each frame.

'timestamps': the elapsed time over each frame, in seconds.

Alternatively, the function can returns a list containing 2 sublists, the first corresponding to the one mentioned above and the second containing all the elements retrieved from the .csv file (see rawDat argument). Also, the function can flip y coordinates (see flipY argument).

Usage

readTrackR(
  trackRPath,
  flipY = FALSE,
  imgHeight = NULL,
  frameR = NULL,
  rawDat = FALSE
)

Arguments

trackRPath

The full path of the TrackR output file (.csv).

flipY

A logical value (i.e., TRUE or FALSE) indicating whether the origin of y coordinates should be flipped. If TRUE, y coordinates are flipped to start on the top-left (default = FALSE).

imgHeight

A numeric value expressed in pixels, the length of Y axis corresponding to the height of the image or video resolution (optional, only used when flipY = TRUE).

frameR

A numeric value expressed in frames per second, the frequency at which frames are recorded/displayed in the video (optional, only used to compute timestamps).

rawDat

TRUE or FALSE, if TRUE add a second list containing all the elements retrieved from .csv file (see trackR), may drastically increase the size of the object returned by the function (default = FALSE).

Value

A list containing either a list of 9 elements classically used for further computations or a list containing 2 sublists, the first corresponding to the one previously mentioned and the second containing all the elements retrieved from the .csv file (see rawDat argument). Also, by default the function returns y coordinates starting on the bottom-left.

References

trackR

See also

Author

Quentin PETITJEAN

Examples

## Not run:

# Download the first dataset from the sample data repository
Path2Data <- MoveR::DLsampleData(dataSet = 1, tracker = "TrackR")
Path2Data
#> [1] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpmYFHq7\\MoveR_SampleData-main\\sample_1\\TrackROutput\\2602_ISA3080_Low_5_Results.csv"                          
#> [2] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpmYFHq7\\MoveR_SampleData-main\\sample_1\\ReferenceData\\DistMatrixFromArenaEdge_2602_ISA3080_Low_5.mov_1800.txt"
#> [3] "C:\\Users\\quent\\AppData\\Local\\Temp\\RtmpmYFHq7\\MoveR_SampleData-main\\sample_1\\ReferenceData\\RefDat_2602_ISA3080_Low_5.csv"                          

# Import the list containing the 9 vectors classically used for further computation
# and flip Y coordinates to start on the top-left
Data <-
  MoveR::readTrackR(Path2Data[[1]],
         flipY = T,
         imgHeight = 2160,
         rawDat = F
  )
#> Warning: frameR argument is missing: timestamps returned NA, a frame rate value is needed to compute timestamps
str(Data)
#> List of 9
#>  $ maj.ax    : num [1:134835] 17 17.7 19.6 14.1 15.3 ...
#>  $ angle     : num [1:134835] 0.689 0.73 2.282 0.294 0.443 ...
#>  $ min.ax    : num [1:134835] 10.1 10 10.8 10.3 10.6 ...
#>  $ x.pos     : num [1:134835] 299 1873.2 183.3 81.5 565.5 ...
#>  $ y.pos     : num [1:134835] 167 287 359 392 453 ...
#>  $ identity  : int [1:134835] 1 2 3 4 5 6 7 8 9 10 ...
#>  $ frame     : int [1:134835] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ ntargets  : int [1:12033] 11 11 12 12 12 12 12 12 11 11 ...
#>  $ timestamps: logi [1:12033] NA NA NA NA NA NA ...

# Import the list containing 2 sublists, the first containing the 9 vectors classically used for further computation
# and the second list containing all the elements retrieved from .csv file,
# also do not flip Y coordinates (start on the bottom-left)

DataFull <-
  MoveR::readTrackR(Path2Data[[1]],
    rawDat = T,
    frameR = 25
  )

str(DataFull)
#> List of 2
#>  $ Data_trackR    :List of 9
#>   ..$ maj.ax    : num [1:134835] 17 17.7 19.6 14.1 15.3 ...
#>   ..$ angle     : num [1:134835] 0.689 0.73 2.282 0.294 0.443 ...
#>   ..$ min.ax    : num [1:134835] 10.1 10 10.8 10.3 10.6 ...
#>   ..$ x.pos     : num [1:134835] 299 1873.2 183.3 81.5 565.5 ...
#>   ..$ y.pos     : num [1:134835] 1993 1873 1801 1768 1707 ...
#>   ..$ identity  : int [1:134835] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ frame     : int [1:134835] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ ntargets  : int [1:12033] 11 11 12 12 12 12 12 12 11 11 ...
#>   ..$ timestamps: num [1:12033] 0.04 0.08 0.12 0.16 0.2 0.24 0.28 0.32 0.36 0.4 ...
#>  $ Data_trackR_Raw:List of 8
#>   ..$ frame : int [1:134835] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ track : int [1:134835] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ x     : num [1:134835] 299 1873.2 183.3 81.5 565.5 ...
#>   ..$ y     : num [1:134835] 1993 1873 1801 1768 1707 ...
#>   ..$ width : num [1:134835] 10.1 10 10.8 10.3 10.6 ...
#>   ..$ height: num [1:134835] 17 17.7 19.6 14.1 15.3 ...
#>   ..$ angle : num [1:134835] 39.5 41.8 130.7 16.8 25.4 ...
#>   ..$ n     : int [1:134835] 135 139 165 113 128 145 153 133 143 158 ...

## End(Not run)